Move CSV column parsers to separate module
